# Implementation/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/allocation/strategy": {
      "get": {
        "description": "R8: la dashboard operatore mostra sempre quale politica sta assegnando i veicoli (DD §3.2). Il valore viene dal record `system_mode`, l'unica sede autorevole.",
        "operationId": "AllocationController_getActiveStrategy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveStrategyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token assente, scaduto o non valido."
          },
          "403": {
            "description": "Serve un account operatore."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Legge la strategia di allocazione attiva",
        "tags": [
          "allocation"
        ]
      },
      "put": {
        "description": "R8: il cambio ha effetto sulla prima allocazione successiva, senza interruzione del servizio. Essendo una scelta umana porta il sistema in modo Manual nella stessa transazione, e da quel momento nessun cambio automatico avviene finché un operatore non riabilita il modo Auto (NFR10, M6).",
        "operationId": "AllocationController_setActiveStrategy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveStrategyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveStrategyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token assente, scaduto o non valido."
          },
          "403": {
            "description": "Serve un account operatore."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambia la strategia di allocazione",
        "tags": [
          "allocation"
        ]
      }
    },
...
    "/mode": {
      "get": {
        "description": "NFR10: il modo corrente è sempre visibile sulla dashboard. I due valori vengono dal record `system_mode`, la loro unica sede autorevole, quindi due repliche del tier applicativo rispondono la stessa cosa (NFR3).",
        "operationId": "ControlModeController_getMode",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token assente, scaduto o non valido."
          },
          "403": {
            "description": "Serve un account operatore."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Legge il modo di controllo e la strategia attiva",
        "tags": [
          "mode"
        ]
      },
      "put": {
        "description": "R13: il sistema resta in Manual finché l'operatore non riabilita Auto esplicitamente. Il rientro **rivaluta subito** l'ultimo livello di traffico noto e applica la strategia che gli compete, quindi la strategia nella risposta può essere diversa da quella scelta a mano. Se il livello è Medium resta quella attiva in quell istante.",
        "operationId": "ControlModeController_enableAuto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableAutoModeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token assente, scaduto o non valido."
          },
          "403": {
            "description": "Serve un account operatore."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Riabilita il modo Auto",
        "tags": [
          "mode"
        ]
      }
    },
...
  },
  "info": {
    "title": "ROAd API",
    "description": "Contratto pubblico del backend ROAd. È la sola cosa che i client devono conoscere: nessun client importa codice da apps/api.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ActiveStrategyResponseDto": {
        "type": "object",
        "properties": {
          "activeStrategy": {
            "type": "string",
            "enum": [
              "NEAREST_AVAILABLE",
              "MINIMUM_ETA"
            ],
            "example": "NEAREST_AVAILABLE",
            "description": "La politica con cui il sistema sta assegnando i veicoli in questo momento."
          }
        },
        "required": [
          "activeStrategy"
        ]
      },
      "SetActiveStrategyRequestDto": {
        "type": "object",
        "properties": {
          "strategy": {
            "type": "string",
            "enum": [
              "NEAREST_AVAILABLE",
              "MINIMUM_ETA"
            ],
            "example": "MINIMUM_ETA",
            "description": "La politica da rendere attiva. La scelta di un operatore è per definizione un intervento manuale: porta il sistema in modo Manual e sospende i cambi automatici (NFR10)."
          }
        },
        "required": [
          "strategy"
        ]
      },
...
      "ModeResponseDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ],
            "example": "AUTO",
            "description": "Il modo di controllo corrente. In Auto il sistema commuta la strategia in base al traffico; in Manual segue la scelta dell operatore e sospende ogni cambio automatico."
          },
          "activeStrategy": {
            "type": "string",
            "enum": [
              "NEAREST_AVAILABLE",
              "MINIMUM_ETA"
            ],
            "example": "NEAREST_AVAILABLE",
            "description": "La politica attiva dopo l operazione. Compare anche qui perché il pannello di controllo della dashboard mostra i due valori insieme, e perché il rientro in modo Auto può cambiarla subito rivalutando l ultimo livello di traffico noto."
          }
        },
        "required": [
          "mode",
          "activeStrategy"
        ]
      },
      "EnableAutoModeRequestDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "AUTO"
            ],
            "example": "AUTO",
            "description": "L unico valore ammesso. In modo Manual non ci si porta dichiarandolo: ci si finisce scegliendo una strategia su PUT /allocation/strategy, come prescrive R13."
          }
        },
        "required": [
          "mode"
        ]
      },
...
    }
  }
}